+++ data/static/js/app.js
@@ -7,6 +7,7 @@
         CT.initWizard();
         CT.initSettings();
         CT.render();
+        CT.reloadEntityIds().then(function () { CT.render(); });
         CT.reloadStatus().then(CT.renderStatus).catch(function () {});
         if (!CT.socket) CT.error('Connexion temps réel indisponible : rechargez la page.');
     }